+++ resolwe_bio/processes/support_processors/seqtk_reverse_complement.py
@@ -23,8 +23,11 @@ class ReverseComplementSingle(Process):
             'memory': 16384,
         },
     }
+    entity = {
+        'type': 'sample',
+    }
     data_name = '{{ reads|sample_name|default("?") }}'
-    version = '1.0.0'
+    version = '1.0.1'
 
     class Input:
         """Input fields to process ReverseComplementSingle."""
@@ -79,8 +82,11 @@ class ReverseComplementPaired(Process):
             'memory': 16384,
         },
     }
+    entity = {
+        'type': 'sample',
+    }
     data_name = '{{ reads|sample_name|default("?") }}'
-    version = '1.0.0'
+    version = '1.0.1'
 
     class Input:
         """Input fields to process ReverseComplementPaired."""